refactor the admin cmd: move the subcmd to differenct files
